requirements.txt
langchain
langgraph
streamlit
python-dotenv==1.0.1
psycopg[binary]
kiwipiepy
konlpy
PyYAML
pyfiglet
pytest